Added Auth Mutation Tests
