+++ server/routes/auth/resolvers.ts
@@ -4,7 +4,7 @@ const AuthType: GraphQLObjectType = new GraphQLObjectType({
 	name: 'AuthType',
 	description: 'Authentication Token Data',
 	fields: () => ({
-		_id: { type: GraphQLNonNull(GraphQLString) },
+		id: { type: GraphQLNonNull(GraphQLString) },
 		avatar: { type: GraphQLNonNull(GraphQLString) },
 		email: { type: GraphQLNonNull(GraphQLString) },
 		name: { type: GraphQLNonNull(GraphQLString) },
